refactor(packages): delete address-mapper, add it to currency package
