--- packages/address-mapper/.eslintrc.js
@@ -1,1 +1,0 @@
-module.exports = require('@sushiswap/eslint-config/nextjs')
