--- packages/address-mapper/.prettierrc.js
@@ -1,3 +1,0 @@
-const { default: config } = require('@sushiswap/prettier-config')
-
-module.exports = config
