--- packages/address-mapper/README.md
@@ -1,3 +1,0 @@
-# Address mapper
-
-...
