--- packages/address-mapper/tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "extends": "@sushiswap/typescript-config/esm.json",
-  "include": ["."],
-  "exclude": ["dist", "node_modules"],
-  "compilerOptions": {
-    "declaration": false,
-    "declarationMap": false,
-    "target":"ES2015"
-   }
-}
