+++ packages/currency/src/__tests__/AddressMapper.test.ts
@@ -1,6 +1,7 @@
 
-import { describe, expect, test } from "vitest";
-import { AddressMapper } from "../src/index.js";
+import { AddressMapper } from "../AddressMapper";
+import { ChainId } from "@sushiswap/chain";
+import { SUSHI_ADDRESS, WNATIVE_ADDRESS } from "../constants";
 
 const USDC_ETHEREUM_BRIDGE_LIST = {
     1: "0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48",
@@ -12,12 +13,9 @@ const USDC_BSC_BRIDGE_LIST = {
     1: "0x0000000000000000000000000000000000000101",
 }
 
-
 describe("AddressMapper", () => {
 
-    
     const usdc_merged = AddressMapper.merge(USDC_ETHEREUM_BRIDGE_LIST, USDC_BSC_BRIDGE_LIST)
-
 
     test("merge", async () => {
         const expected = {
@@ -27,7 +25,6 @@ describe("AddressMapper", () => {
         }
         expect(usdc_merged).toEqual(expected)
     });
-
 
     test("when kovan address is given, return all eth and ropsten addresses", async () => {
         const usdc_merged = AddressMapper.merge(USDC_ETHEREUM_BRIDGE_LIST, USDC_BSC_BRIDGE_LIST)
@@ -81,7 +78,6 @@ describe("AddressMapper", () => {
             expect(actual).toEqual(expected)
         })
 
-
     describe("case sensitivity", () => {
         const testList = {
             1: "0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48",
@@ -93,10 +89,40 @@ describe("AddressMapper", () => {
             expect(mappings['1:0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48']).toEqual({ 4: ["0x1717a0d5c8705ee89a8ad6e808268d6a826c97a4"] })
         })
 
-        
         test("returns undefined when key isn't lower case", async () => {
             expect(mappings['1:0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48']).toBeUndefined()
         })
     })
     
+    describe("Sushi addresses", () => {
+        const addressLists = [
+            WNATIVE_ADDRESS,
+            SUSHI_ADDRESS,
+        ]
+        const MAPPING = AddressMapper.generate([...addressLists]);
+
+        test("Get all weth addresses", async () => {
+            const expected = Object.keys(WNATIVE_ADDRESS).length - 1 // minus one, exclude itself 
+            const actual = Object.values(MAPPING[`${ChainId.ETHEREUM}:${WNATIVE_ADDRESS[ChainId.ETHEREUM].toLowerCase()}`]).length
+            expect(actual).toEqual(expected);
+        });
+        
+        test("Get all sushi addresses for the top networks given thundercores sushi address", async () => {
+
+            const actual = MAPPING[`${ChainId.THUNDERCORE}:${SUSHI_ADDRESS[ChainId.THUNDERCORE].toLowerCase()}`]
+
+            expect(actual[ChainId.ETHEREUM]).toHaveLength(1);
+            expect(actual[ChainId.ETHEREUM][0]).toEqual(SUSHI_ADDRESS[ChainId.ETHEREUM].toLowerCase());
+            
+            expect(actual[ChainId.ARBITRUM]).toHaveLength(1);
+            expect(actual[ChainId.ARBITRUM][0]).toEqual(SUSHI_ADDRESS[ChainId.ARBITRUM].toLowerCase());
+            
+            expect(actual[ChainId.POLYGON]).toHaveLength(1);
+            expect(actual[ChainId.POLYGON][0]).toEqual(SUSHI_ADDRESS[ChainId.POLYGON].toLowerCase());
+            
+            expect(actual[ChainId.OPTIMISM]).toHaveLength(1);
+            expect(actual[ChainId.OPTIMISM][0]).toEqual(SUSHI_ADDRESS[ChainId.OPTIMISM].toLowerCase());
+        });
+    });
 });
+
